+++ docs/html/search/all_0.js
@@ -3,7 +3,7 @@ var searchData=
   ['abbrev_0',['abbrev',['../structace__time_1_1basic_1_1TransitionTemplate.html#abe2b5dd9d3dd7660c10e85a388494c2b',1,'ace_time::basic::TransitionTemplate::abbrev()'],['../structace__time_1_1extended_1_1TransitionTemplate.html#ad65c299286c1f8a96ff214ccddffd3f1',1,'ace_time::extended::TransitionTemplate::abbrev()'],['../classace__time_1_1FindResult.html#adf56ea188f29e6dde228ba2bdf70d05e',1,'ace_time::FindResult::abbrev()'],['../classace__time_1_1ZonedExtra.html#a8c05778e4b263ae9a46232ec5b8a6bd3',1,'ace_time::ZonedExtra::abbrev()']]],
   ['ace_5ftime_5fuse_5fprogmem_1',['ACE_TIME_USE_PROGMEM',['../compat_8h.html#a736a584901fd7e4d3c06ade513afd259',1,'compat.h']]],
   ['acetime_20library_2',['AceTime Library',['../index.html',1,'']]],
-  ['acetime_5ft_3',['acetime_t',['../common_2common_8h.html#a9c5a9893b2588031b6ab56826dec303a',1,'ace_time']]],
+  ['acetime_5ft_3',['acetime_t',['../common_8h.html#a9c5a9893b2588031b6ab56826dec303a',1,'ace_time']]],
   ['addactivecandidatestoactivepool_4',['addActiveCandidatesToActivePool',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html#aaa860da6313337525d7fe83f532994b0',1,'ace_time::extended::TransitionStorageTemplate']]],
   ['addfreeagenttoactivepool_5',['addFreeAgentToActivePool',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html#a0b23350e74c9e433d6d7d76111aeccfa',1,'ace_time::extended::TransitionStorageTemplate']]],
   ['addfreeagenttocandidatepool_6',['addFreeAgentToCandidatePool',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html#a7a242ca931d608e3913612fc662bb0e2',1,'ace_time::extended::TransitionStorageTemplate']]],