+++ docs/html/search/classes_0.js
@@ -1,10 +1,11 @@
 var searchData=
 [
-  ['basiczone_315',['BasicZone',['../classace__time_1_1BasicZone.html',1,'ace_time']]],
-  ['basiczonemanager_316',['BasicZoneManager',['../classace__time_1_1BasicZoneManager.html',1,'ace_time']]],
-  ['basiczoneprocessor_317',['BasicZoneProcessor',['../classace__time_1_1BasicZoneProcessor.html',1,'ace_time']]],
-  ['basiczoneprocessorcache_318',['BasicZoneProcessorCache',['../classace__time_1_1BasicZoneProcessorCache.html',1,'ace_time']]],
-  ['basiczoneprocessortemplate_319',['BasicZoneProcessorTemplate',['../classace__time_1_1BasicZoneProcessorTemplate.html',1,'ace_time']]],
-  ['basiczoneprocessortemplate_3c_20basic_3a_3abrokerfactory_2c_20basic_3a_3azoneinfobroker_2c_20basic_3a_3azoneerabroker_2c_20basic_3a_3azonepolicybroker_2c_20basic_3a_3azonerulebroker_20_3e_320',['BasicZoneProcessorTemplate&lt; basic::BrokerFactory, basic::ZoneInfoBroker, basic::ZoneEraBroker, basic::ZonePolicyBroker, basic::ZoneRuleBroker &gt;',['../classace__time_1_1BasicZoneProcessorTemplate.html',1,'ace_time']]],
-  ['brokerfactory_321',['BrokerFactory',['../classace__time_1_1basic_1_1BrokerFactory.html',1,'ace_time::basic::BrokerFactory'],['../classace__time_1_1extended_1_1BrokerFactory.html',1,'ace_time::extended::BrokerFactory']]]
+  ['basiczone_319',['BasicZone',['../classace__time_1_1BasicZone.html',1,'ace_time']]],
+  ['basiczonemanager_320',['BasicZoneManager',['../classace__time_1_1BasicZoneManager.html',1,'ace_time']]],
+  ['basiczoneprocessor_321',['BasicZoneProcessor',['../classace__time_1_1BasicZoneProcessor.html',1,'ace_time']]],
+  ['basiczoneprocessorcache_322',['BasicZoneProcessorCache',['../classace__time_1_1BasicZoneProcessorCache.html',1,'ace_time']]],
+  ['basiczoneprocessortemplate_323',['BasicZoneProcessorTemplate',['../classace__time_1_1BasicZoneProcessorTemplate.html',1,'ace_time']]],
+  ['basiczoneprocessortemplate_3c_20basic_3a_3abrokerfactory_2c_20basic_3a_3azoneinfobroker_2c_20basic_3a_3azoneerabroker_2c_20basic_3a_3azonepolicybroker_2c_20basic_3a_3azonerulebroker_20_3e_324',['BasicZoneProcessorTemplate&lt; basic::BrokerFactory, basic::ZoneInfoBroker, basic::ZoneEraBroker, basic::ZonePolicyBroker, basic::ZoneRuleBroker &gt;',['../classace__time_1_1BasicZoneProcessorTemplate.html',1,'ace_time']]],
+  ['brokerfactory_325',['BrokerFactory',['../classace__time_1_1internal_1_1BrokerFactory.html',1,'ace_time::internal']]],
+  ['brokerfactory_3c_20zoneinfo_2c_20zoneera_2c_20zonepolicy_2c_20zonerule_20_3e_326',['BrokerFactory&lt; ZoneInfo, ZoneEra, ZonePolicy, ZoneRule &gt;',['../classace__time_1_1internal_1_1BrokerFactory.html',1,'ace_time::internal']]]
 ];