+++ docs/html/search/classes_7.js
@@ -1,12 +1,12 @@
 var searchData=
 [
-  ['timeoffset_341',['TimeOffset',['../classace__time_1_1TimeOffset.html',1,'ace_time']]],
-  ['timeperiod_342',['TimePeriod',['../classace__time_1_1TimePeriod.html',1,'ace_time']]],
-  ['timezone_343',['TimeZone',['../classace__time_1_1TimeZone.html',1,'ace_time']]],
-  ['timezonedata_344',['TimeZoneData',['../structace__time_1_1TimeZoneData.html',1,'ace_time']]],
-  ['transitionfordatetimetemplate_345',['TransitionForDateTimeTemplate',['../structace__time_1_1extended_1_1TransitionForDateTimeTemplate.html',1,'ace_time::extended']]],
-  ['transitionforsecondstemplate_346',['TransitionForSecondsTemplate',['../structace__time_1_1extended_1_1TransitionForSecondsTemplate.html',1,'ace_time::extended']]],
-  ['transitionstoragetemplate_347',['TransitionStorageTemplate',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html',1,'ace_time::extended']]],
-  ['transitionstoragetemplate_3c_20kmaxtransitions_2c_20zeb_2c_20zpb_2c_20zrb_20_3e_348',['TransitionStorageTemplate&lt; kMaxTransitions, ZEB, ZPB, ZRB &gt;',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html',1,'ace_time::extended']]],
-  ['transitiontemplate_349',['TransitionTemplate',['../structace__time_1_1basic_1_1TransitionTemplate.html',1,'ace_time::basic::TransitionTemplate&lt; ZIB, ZEB, ZPB, ZRB &gt;'],['../structace__time_1_1extended_1_1TransitionTemplate.html',1,'ace_time::extended::TransitionTemplate&lt; ZEB, ZPB, ZRB &gt;']]]
+  ['timeoffset_346',['TimeOffset',['../classace__time_1_1TimeOffset.html',1,'ace_time']]],
+  ['timeperiod_347',['TimePeriod',['../classace__time_1_1TimePeriod.html',1,'ace_time']]],
+  ['timezone_348',['TimeZone',['../classace__time_1_1TimeZone.html',1,'ace_time']]],
+  ['timezonedata_349',['TimeZoneData',['../structace__time_1_1TimeZoneData.html',1,'ace_time']]],
+  ['transitionfordatetimetemplate_350',['TransitionForDateTimeTemplate',['../structace__time_1_1extended_1_1TransitionForDateTimeTemplate.html',1,'ace_time::extended']]],
+  ['transitionforsecondstemplate_351',['TransitionForSecondsTemplate',['../structace__time_1_1extended_1_1TransitionForSecondsTemplate.html',1,'ace_time::extended']]],
+  ['transitionstoragetemplate_352',['TransitionStorageTemplate',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html',1,'ace_time::extended']]],
+  ['transitionstoragetemplate_3c_20kmaxtransitions_2c_20zeb_2c_20zpb_2c_20zrb_20_3e_353',['TransitionStorageTemplate&lt; kMaxTransitions, ZEB, ZPB, ZRB &gt;',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html',1,'ace_time::extended']]],
+  ['transitiontemplate_354',['TransitionTemplate',['../structace__time_1_1basic_1_1TransitionTemplate.html',1,'ace_time::basic::TransitionTemplate&lt; ZIB, ZEB, ZPB, ZRB &gt;'],['../structace__time_1_1extended_1_1TransitionTemplate.html',1,'ace_time::extended::TransitionTemplate&lt; ZEB, ZPB, ZRB &gt;']]]
 ];