+++ docs/html/search/files_3.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['offset_5fdate_5ftime_5fmutation_2eh_376',['offset_date_time_mutation.h',['../offset__date__time__mutation_8h.html',1,'']]]
+  ['time_5foffset_5fmutation_2eh_382',['time_offset_mutation.h',['../time__offset__mutation_8h.html',1,'']]],
+  ['time_5fperiod_5fmutation_2eh_383',['time_period_mutation.h',['../time__period__mutation_8h.html',1,'']]]
 ];