+++ docs/html/search/functions_0.js
@@ -1,8 +1,8 @@
 var searchData=
 [
-  ['abbrev_380',['abbrev',['../classace__time_1_1ZonedExtra.html#a8c05778e4b263ae9a46232ec5b8a6bd3',1,'ace_time::ZonedExtra']]],
-  ['addactivecandidatestoactivepool_381',['addActiveCandidatesToActivePool',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html#aaa860da6313337525d7fe83f532994b0',1,'ace_time::extended::TransitionStorageTemplate']]],
-  ['addfreeagenttoactivepool_382',['addFreeAgentToActivePool',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html#a0b23350e74c9e433d6d7d76111aeccfa',1,'ace_time::extended::TransitionStorageTemplate']]],
-  ['addfreeagenttocandidatepool_383',['addFreeAgentToCandidatePool',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html#a7a242ca931d608e3913612fc662bb0e2',1,'ace_time::extended::TransitionStorageTemplate']]],
-  ['addpriortocandidatepool_384',['addPriorToCandidatePool',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html#ad098d5eb67b3ae5f4fbeda5c5e4a1470',1,'ace_time::extended::TransitionStorageTemplate']]]
+  ['abbrev_385',['abbrev',['../classace__time_1_1ZonedExtra.html#a8c05778e4b263ae9a46232ec5b8a6bd3',1,'ace_time::ZonedExtra']]],
+  ['addactivecandidatestoactivepool_386',['addActiveCandidatesToActivePool',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html#aaa860da6313337525d7fe83f532994b0',1,'ace_time::extended::TransitionStorageTemplate']]],
+  ['addfreeagenttoactivepool_387',['addFreeAgentToActivePool',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html#a0b23350e74c9e433d6d7d76111aeccfa',1,'ace_time::extended::TransitionStorageTemplate']]],
+  ['addfreeagenttocandidatepool_388',['addFreeAgentToCandidatePool',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html#a7a242ca931d608e3913612fc662bb0e2',1,'ace_time::extended::TransitionStorageTemplate']]],
+  ['addpriortocandidatepool_389',['addPriorToCandidatePool',['../classace__time_1_1extended_1_1TransitionStorageTemplate.html#ad098d5eb67b3ae5f4fbeda5c5e4a1470',1,'ace_time::extended::TransitionStorageTemplate']]]
 ];