+++ docs/html/search/typedefs_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['matchingera_645',['MatchingEra',['../classace__time_1_1ExtendedZoneProcessorTemplate.html#afd5f76dcd187f20cb708e7c090c42139',1,'ace_time::ExtendedZoneProcessorTemplate']]]
+  ['matchingera_651',['MatchingEra',['../classace__time_1_1ExtendedZoneProcessorTemplate.html#afd5f76dcd187f20cb708e7c090c42139',1,'ace_time::ExtendedZoneProcessorTemplate']]]
 ];