+++ docs/html/search/variables_2.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['deltaminutes_569',['deltaMinutes',['../structace__time_1_1basic_1_1TransitionTemplate.html#aefb15f684d43cd60da7511701b939a8f',1,'ace_time::basic::TransitionTemplate::deltaMinutes()'],['../structace__time_1_1extended_1_1TransitionTemplate.html#ae9e5b29eca212f58b3af5d3b8cd39351',1,'ace_time::extended::TransitionTemplate::deltaMinutes()']]],
-  ['dstoffsetminutes_570',['dstOffsetMinutes',['../classace__time_1_1FindResult.html#a4bcce5a5c8039f6669fcc91b61926e3d',1,'ace_time::FindResult']]]
+  ['deltaminutes_572',['deltaMinutes',['../structace__time_1_1basic_1_1TransitionTemplate.html#aefb15f684d43cd60da7511701b939a8f',1,'ace_time::basic::TransitionTemplate::deltaMinutes()'],['../structace__time_1_1extended_1_1TransitionTemplate.html#ae9e5b29eca212f58b3af5d3b8cd39351',1,'ace_time::extended::TransitionTemplate::deltaMinutes()']]],
+  ['dstoffsetminutes_573',['dstOffsetMinutes',['../classace__time_1_1FindResult.html#a4bcce5a5c8039f6669fcc91b61926e3d',1,'ace_time::FindResult']]]
 ];