+++ docs/html/search/variables_8.js
@@ -1,8 +1,8 @@
 var searchData=
 [
-  ['match_618',['match',['../structace__time_1_1extended_1_1TransitionTemplate.html#af39b0cf9cf43ddd39feecec6ca483a20',1,'ace_time::extended::TransitionTemplate']]],
-  ['matchstatus_619',['matchStatus',['../structace__time_1_1extended_1_1TransitionTemplate.html#ac9ff2536c93fd1810a90f1bc34655847',1,'ace_time::extended::TransitionTemplate']]],
-  ['month_620',['month',['../structace__time_1_1basic_1_1TransitionTemplate.html#a10415486cabd71d413f6ff942abb938b',1,'ace_time::basic::TransitionTemplate']]],
-  ['mzonekey_621',['mZoneKey',['../classace__time_1_1TimeZone.html#a8eafcad4aa772eb9ea6cbee9ba7110d8',1,'ace_time::TimeZone']]],
-  ['mzoneprocessor_622',['mZoneProcessor',['../classace__time_1_1TimeZone.html#a3380618145722bb6a76108db95153e42',1,'ace_time::TimeZone']]]
+  ['match_623',['match',['../structace__time_1_1extended_1_1TransitionTemplate.html#af39b0cf9cf43ddd39feecec6ca483a20',1,'ace_time::extended::TransitionTemplate']]],
+  ['maxtransitions_624',['maxTransitions',['../structace__time_1_1internal_1_1ZoneContext.html#ad80e6f8f245fa2892ad3551459ff00ae',1,'ace_time::internal::ZoneContext']]],
+  ['month_625',['month',['../structace__time_1_1basic_1_1TransitionTemplate.html#a10415486cabd71d413f6ff942abb938b',1,'ace_time::basic::TransitionTemplate']]],
+  ['mzonekey_626',['mZoneKey',['../classace__time_1_1TimeZone.html#a8eafcad4aa772eb9ea6cbee9ba7110d8',1,'ace_time::TimeZone']]],
+  ['mzoneprocessor_627',['mZoneProcessor',['../classace__time_1_1TimeZone.html#a3380618145722bb6a76108db95153e42',1,'ace_time::TimeZone']]]
 ];