+++ docs/html/search/variables_c.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['reqdstoffsetminutes_628',['reqDstOffsetMinutes',['../classace__time_1_1FindResult.html#ab306e11a047e780cd345a21f1edda280',1,'ace_time::FindResult']]],
-  ['reqstdoffsetminutes_629',['reqStdOffsetMinutes',['../classace__time_1_1FindResult.html#a8b85a4eb01b435551b0f2298318cd3db',1,'ace_time::FindResult']]],
-  ['rule_630',['rule',['../structace__time_1_1basic_1_1TransitionTemplate.html#a2ce8f0d17f189a25a6f80e9107edb239',1,'ace_time::basic::TransitionTemplate::rule()'],['../structace__time_1_1extended_1_1TransitionTemplate.html#a08e0996372e9c0c80fc60e9a5ceff41e',1,'ace_time::extended::TransitionTemplate::rule()']]]
+  ['reqdstoffsetminutes_634',['reqDstOffsetMinutes',['../classace__time_1_1FindResult.html#ab306e11a047e780cd345a21f1edda280',1,'ace_time::FindResult']]],
+  ['reqstdoffsetminutes_635',['reqStdOffsetMinutes',['../classace__time_1_1FindResult.html#a8b85a4eb01b435551b0f2298318cd3db',1,'ace_time::FindResult']]],
+  ['rule_636',['rule',['../structace__time_1_1basic_1_1TransitionTemplate.html#a2ce8f0d17f189a25a6f80e9107edb239',1,'ace_time::basic::TransitionTemplate']]]
 ];