+++ src/ace_time/Transition.h
@@ -6,9 +6,10 @@
 #ifndef ACE_TIME_EXTENDED_TRANSITION_H
 #define ACE_TIME_EXTENDED_TRANSITION_H
 
-#include <stdint.h> // uintptr_t
+#include <stdint.h> // uint8_t
 #include "common/logging.h"
 #include "local_date_mutation.h"
+#include "DateTuple.h"
 
 class TransitionStorageTest_getFreeAgent;
 class TransitionStorageTest_getFreeAgent2;
@@ -30,158 +31,11 @@ class Print;
 namespace ace_time {
 namespace extended {
 
-//---------------------------------------------------------------------------
-
-/**
- * The result of comparing the transition time of a Transition to the
- * time interval of its corresponding MatchingEra.
- */
-enum class MatchStatus : uint8_t {
-  kFarPast, // 0
-  kPrior, // 1
-  kExactMatch, // 2
-  kWithinMatch, // 3
-  kFarFuture, // 4
-};
-
-inline bool isMatchStatusActive(MatchStatus status) {
-  return status == MatchStatus::kExactMatch
-      || status == MatchStatus::kWithinMatch
-      || status == MatchStatus::kPrior;
+inline bool isCompareStatusActive(CompareStatus status) {
+  return status == CompareStatus::kExactMatch
+      || status == CompareStatus::kWithinMatch
+      || status == CompareStatus::kPrior;
 }
-
-//---------------------------------------------------------------------------
-
-/**
- * A tuple that represents a date and time. Packed to 4-byte boundaries to
- * save space on 32-bit processors.
- */
-struct DateTuple {
-  DateTuple() = default;
-
-  DateTuple(int16_t y, uint8_t mon, uint8_t d, int16_t min, uint8_t mod):
-      year(y), month(mon), day(d), suffix(mod), minutes(min) {}
-
-  int16_t year; // [-1,10000]
-  uint8_t month; // [1,12]
-  uint8_t day; // [1,31]
-  uint8_t suffix; // kSuffixS, kSuffixW, kSuffixU
-  int16_t minutes; // negative values allowed
-
-  /** Used only for debugging. */
-  void log() const {
-    if (ACE_TIME_EXTENDED_ZONE_PROCESSOR_DEBUG) {
-      int hour = minutes / 60;
-      int minute = minutes - hour * 60;
-      char c = "wsu"[(suffix>>4)];
-      logging::printf("%04d-%02u-%02uT%02d:%02d%c",
-          year, month, day, hour, minute, c);
-    }
-  }
-};
-
-/** Determine if DateTuple a is less than DateTuple b, ignoring the suffix. */
-inline bool operator<(const DateTuple& a, const DateTuple& b) {
-  if (a.year < b.year) return true;
-  if (a.year > b.year) return false;
-  if (a.month < b.month) return true;
-  if (a.month > b.month) return false;
-  if (a.day < b.day) return true;
-  if (a.day > b.day) return false;
-  if (a.minutes < b.minutes) return true;
-  if (a.minutes > b.minutes) return false;
-  return false;
-}
-
-inline bool operator>=(const DateTuple& a, const DateTuple& b) {
-  return ! (a < b);
-}
-
-inline bool operator<=(const DateTuple& a, const DateTuple& b) {
-  return ! (b < a);
-}
-
-inline bool operator>(const DateTuple& a, const DateTuple& b) {
-  return (b < a);
-}
-
-/** Determine if DateTuple a is equal to DateTuple b, including the suffix. */
-inline bool operator==(const DateTuple& a, const DateTuple& b) {
-  return a.year == b.year
-      && a.month == b.month
-      && a.day == b.day
-      && a.minutes == b.minutes
-      && a.suffix == b.suffix;
-}
-
-/** Normalize DateTuple::minutes if its magnitude is more than 24 hours. */
-inline void normalizeDateTuple(DateTuple* dt) {
-  const int16_t kOneDayAsMinutes = 60 * 24;
-  if (dt->minutes <= -kOneDayAsMinutes) {
-    LocalDate ld = LocalDate::forComponents(dt->year, dt->month, dt->day);
-    local_date_mutation::decrementOneDay(ld);
-    dt->year = ld.year();
-    dt->month = ld.month();
-    dt->day = ld.day();
-    dt->minutes += kOneDayAsMinutes;
-  } else if (kOneDayAsMinutes <= dt->minutes) {
-    LocalDate ld = LocalDate::forComponents(dt->year, dt->month, dt->day);
-    local_date_mutation::incrementOneDay(ld);
-    dt->year = ld.year();
-    dt->month = ld.month();
-    dt->day = ld.day();
-    dt->minutes -= kOneDayAsMinutes;
-  } else {
-    // do nothing
-  }
-}
-
-/**
- * Return the number of seconds in (a - b), ignoring suffix. This function is
- * valid for all years [1, 10000), regardless of the Epoch::currentEpochYear(),
- * as long as the difference between the two DateTuples fits inside an
- * `acetime_t`, which is a signed 32-bit integer.
- */
-inline acetime_t subtractDateTuple(const DateTuple& a, const DateTuple& b) {
-  int32_t epochDaysA = LocalDate::forComponents(
-      a.year, a.month, a.day).toEpochDays();
-
-  int32_t epochDaysB = LocalDate::forComponents(
-      b.year, b.month, b.day).toEpochDays();
-
-  // Perform the subtraction of the days first, before converting to seconds, to
-  // prevent overflow if a.year or b.year is more than 68 years from the current
-  // epoch year.
-  return (epochDaysA - epochDaysB) * 86400 + (a.minutes - b.minutes) * 60;
-}
-
-/**
-  * Determine the relationship of t to the time interval defined by `[start,
-  * until)`. The comparison is fuzzy, with a slop of about one month so that
-  * we can ignore the day and minutes fields.
-  *
-  * The following values are returned:
-  *
-  *  * kPrior if 't' is less than 'start' by at least one month,
-  *  * kFarFuture if 't' is greater than 'until' by at least one month,
-  *  * kWithinMatch if 't' is within [start, until) with a one month slop,
-  *  * kExactMatch is never returned.
-  */
-inline MatchStatus compareDateTupleFuzzy(
-    const DateTuple& t,
-    const DateTuple& start,
-    const DateTuple& until) {
-  // Use int32_t because a delta year of 2730 or greater will exceed
-  // the range of an int16_t.
-  int32_t tMonths = t.year * (int32_t) 12 + t.month;
-  int32_t startMonths = start.year * (int32_t) 12 + start.month;
-  if (tMonths < startMonths - 1) return MatchStatus::kPrior;
-  int32_t untilMonths = until.year * 12 + until.month;
-  if (untilMonths + 1 < tMonths) return MatchStatus::kFarFuture;
-  return MatchStatus::kWithinMatch;
-}
-
-//---------------------------------------------------------------------------
 
 /**
  * Data structure that captures the matching ZoneEra and its ZoneRule
@@ -239,7 +93,7 @@ struct MatchingEraTemplate {
  *  given by match->offsetMinutes. The additional DST delta is given by
  *  rule->deltaMinutes.
  *
- * Some of the instance variables (e.g. 'isValidPrior', 'matchStatus',
+ * Some of the instance variables (e.g. 'isValidPrior', 'compareStatus',
  * 'transitionTime', 'transitionTimeS', 'transitionTimeU', 'letter()' and
  * 'format()') are transient parameters which are in the implementation of the
  * TransitionStorage::init() method.
@@ -262,12 +116,15 @@ struct TransitionTemplate {
   /** The match which generated this Transition. */
   const MatchingEraTemplate<ZEB>* match;
 
+#if ACE_TIME_EXTENDED_ZONE_PROCESSOR_DEBUG
   /**
    * The Zone transition rule that matched for the the given year. Set to
    * nullptr if the RULES column is '-', indicating that the MatchingEra was
-   * a "simple" ZoneEra.
+   * a "simple" ZoneEra. This not required for actual calculation, but it is
+   * useful to have a reference to it for debugging purposes.
    */
   ZRB rule;
+#endif
 
   /**
    * The original transition time, usually 'w' but sometimes 's' or 'u'. After
@@ -333,8 +190,8 @@ struct TransitionTemplate {
   /** The calculated effective time zone abbreviation, e.g. "PST" or "PDT". */
   char abbrev[internal::kAbbrevSize];
 
-  /** Storage for the single letter 'letter' field if 'rule' is not null. */
-  char letterBuf[2];
+  /** Storage for the 'letter' field if 'rule' is not null. */
+  const char* letter;
 
   union {
     /**
@@ -346,51 +203,14 @@ struct TransitionTemplate {
     bool isValidPrior;
 
     /**
-     * During processTransitionMatchStatus(), this flag indicates how the
+     * During processTransitionCompareStatus(), this flag indicates how the
      * transition falls within the time interval of the MatchingEra.
      */
-    MatchStatus matchStatus;
+    CompareStatus compareStatus;
   };
-
-  //-------------------------------------------------------------------------
 
   const char* format() const {
     return match->era.format();
-  }
-
-  /**
-   * Return the letter string. Returns nullptr if the RULES column is empty
-   * since that means that the ZoneRule is not used, which means LETTER does
-   * not exist. A LETTER of '-' is returned as an empty string "".
-   */
-  const char* letter() const {
-    // RULES column is '-' or hh:mm, so return nullptr to indicate this.
-    if (rule.isNull()) {
-      return nullptr;
-    }
-
-    // RULES point to a named rule, and LETTER is a single, printable character.
-    // Return the letterBuf which contains a NUL-terminated string containing
-    // the single character, as initialized in createTransitionForYear().
-    char letter = rule.letter();
-    if (letter >= 32) {
-      return letterBuf;
-    }
-
-    // RULES points to a named rule, and the LETTER is a string. The
-    // rule->letter is a non-printable number < 32, which is an index into
-    // a list of strings given by match->era->zonePolicy->letters[].
-    const ZPB policy = match->era.zonePolicy();
-    uint8_t numLetters = policy.numLetters();
-    if (letter >= numLetters) {
-      // This should never happen unless there is a programming error. If it
-      // does, return an empty string. (createTransitionForYear() sets
-      // letterBuf to a NUL terminated empty string if rule->letter < 32)
-      return letterBuf;
-    }
-
-    // Return the string at index 'rule->letter'.
-    return policy.letter(letter);
   }
 
   /** Used only for debugging. */
@@ -401,19 +221,21 @@ struct TransitionTemplate {
     } else {
       logging::printf("start=%ld", startEpochSeconds);
     }
-    logging::printf("; status=%d", matchStatus);
+    logging::printf("; status=%d", compareStatus);
     logging::printf("; UTC");
     logHourMinute(offsetMinutes);
     logHourMinute(deltaMinutes);
     logging::printf("; tt="); transitionTime.log();
     logging::printf("; tts="); transitionTimeS.log();
     logging::printf("; ttu="); transitionTimeU.log();
+  #if ACE_TIME_EXTENDED_ZONE_PROCESSOR_DEBUG
     if (rule.isNull()) {
       logging::printf("; rule=-");
     } else {
       logging::printf("; rule=");
       logging::printf("[%d,%d]", rule.fromYear(), rule.toYear());
     }
+  #endif
   }
 
   /** Print minutes as [+/-]hh:mm. */
@@ -712,7 +534,7 @@ class TransitionStorageTemplate {
       uint8_t iActive = mIndexPrior;
       uint8_t iCandidate = mIndexCandidates;
       for (; iCandidate < mIndexFree; iCandidate++) {
-        if (isMatchStatusActive(mTransitions[iCandidate]->matchStatus)) {
+        if (isCompareStatusActive(mTransitions[iCandidate]->compareStatus)) {
           if (iActive != iCandidate) {
             // Must use swap(), because we are moving pointers instead of the
             // actual Transition objects.